Daily-trends-with-Tor sessions/requirements.txt
stem
twine==1.14.0
pip==21.1
wheel==0.33.6
pandas>=0.24.2
numpy>=1.16.4
requests>=2.22.0
tqdm>=4.35.0
extproxy>=1.0.1
google-cloud-storage
google-api-python-client
